Move the assessment to a descriptive folder
