+++ js/assessments/readability/sentenceLengthInTextAssessment.js
@@ -1,11 +1,11 @@
-var AssessmentResult = require( "../values/AssessmentResult.js" );
-var countTooLongSentences = require( "./../assessmentHelpers/checkForTooLongSentences.js" );
-var formatNumber = require( "../helpers/formatNumber.js" );
-var inRange = require( "../helpers/inRange.js" ).inRangeEndInclusive;
-var stripTags = require( "../stringProcessing/stripHTMLTags" ).stripIncompleteTags;
+var AssessmentResult = require( "../../values/AssessmentResult.js" );
+var countTooLongSentences = require( "./../../assessmentHelpers/checkForTooLongSentences.js" );
+var formatNumber = require( "../../helpers/formatNumber.js" );
+var inRange = require( "../../helpers/inRange.js" ).inRangeEndInclusive;
+var stripTags = require( "../../stringProcessing/stripHTMLTags" ).stripIncompleteTags;
 
-var Mark = require( "../values/Mark.js" );
-var addMark = require( "../markers/addMark.js" );
+var Mark = require( "../../values/Mark.js" );
+var addMark = require( "../../markers/addMark.js" );
 
 var map = require( "lodash/map" );
