+++ js/assessments/seo/metaDescriptionKeywordAssessment.js
@@ -1,4 +1,4 @@
-var AssessmentResult = require( "../values/AssessmentResult.js" );
+var AssessmentResult = require( "../../values/AssessmentResult.js" );
 
 /**
  * Returns the score and text for the description keyword match.
