+++ js/assessments/seo/subheadingsKeywordAssessment.js
@@ -1,4 +1,4 @@
-var AssessmentResult = require( "../values/AssessmentResult.js" );
+var AssessmentResult = require( "../../values/AssessmentResult.js" );
 
 /**
  * Returns a score and text based on the keyword matches object.
