+++ js/assessments/seo/titleKeywordAssessment.js
@@ -1,4 +1,4 @@
-var AssessmentResult = require( "../values/AssessmentResult.js" );
+var AssessmentResult = require( "../../values/AssessmentResult.js" );
 var escape = require( "lodash/escape" );
 
 /**
