+++ spec/assessments/taxonomyTextLengthSpec.js
@@ -1,4 +1,4 @@
-var taxonomyTextLengthAssessment = require( "../../js/assessments/taxonomyTextLengthAssessment.js" );
+var taxonomyTextLengthAssessment = require( "../../js/assessments/seo/taxonomyTextLengthAssessment.js" );
 var Paper = require( "../../js/values/Paper.js" );
 var Factory = require( "../helpers/factory.js" );
 var i18n = Factory.buildJed();
