+++ spec/assessments/titleKeywordSpec.js
@@ -1,4 +1,4 @@
-var pageTitleKeywordAssessment = require( "../../js/assessments/titleKeywordAssessment.js" );
+var pageTitleKeywordAssessment = require( "../../js/assessments/seo/titleKeywordAssessment.js" );
 var Paper = require( "../../js/values/Paper.js" );
 var Factory = require( "../helpers/factory.js" );
 var i18n = Factory.buildJed();
